+++ src/main/java/io/swagger/api/UsersApiController.java
@@ -2,7 +2,6 @@ package io.swagger.api;
 
 import io.swagger.model.Account;
 import java.math.BigDecimal;
-import io.swagger.model.RegistrationUser;
 import io.swagger.model.User;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.annotations.*;
@@ -11,20 +10,15 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.multipart.MultipartFile;
 
-import javax.validation.constraints.*;
 import javax.validation.Valid;
 import javax.servlet.http.HttpServletRequest;
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
+
 @javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-05-18T19:26:09.389Z[GMT]")
 @Controller
 public class UsersApiController implements UsersApi {
@@ -72,10 +66,10 @@ public class UsersApiController implements UsersApi {
     }
 
     public ResponseEntity<List<User>> getUsers(@ApiParam(value = "The number of items to skip before starting to collect the result set") @Valid @RequestParam(value = "offset", required = false) Integer offset
-,@ApiParam(value = "The numbers of items to return") @Valid @RequestParam(value = "limit", required = false) Integer limit
-,@ApiParam(value = "First name of the user") @Valid @RequestParam(value = "firstName", required = false) String firstName
-,@ApiParam(value = "Last name of the user") @Valid @RequestParam(value = "lastName", required = false) String lastName
-,@ApiParam(value = "Access level of this user") @Valid @RequestParam(value = "accessLevel", required = false) BigDecimal accessLevel
+, @ApiParam(value = "The numbers of items to return") @Valid @RequestParam(value = "limit", required = false) Integer limit
+, @ApiParam(value = "First name of the user") @Valid @RequestParam(value = "firstName", required = false) String firstName
+, @ApiParam(value = "Last name of the user") @Valid @RequestParam(value = "lastName", required = false) String lastName
+, @ApiParam(value = "Access level of this user") @Valid @RequestParam(value = "accessLevel", required = false) BigDecimal accessLevel
 ) {
         String accept = request.getHeader("Accept");
         if (accept != null && accept.contains("application/json")) {
@@ -91,7 +85,7 @@ public class UsersApiController implements UsersApi {
     }
 
     public ResponseEntity<User> registerUser(@ApiParam(value = "User object to register to the database") @Valid @RequestParam(value = "firstName", required = false) String firstName
-,@ApiParam(value = "User object to register to the database") @Valid @RequestParam(value = "lastName", required = false) String lastName
+, @ApiParam(value = "User object to register to the database") @Valid @RequestParam(value = "lastName", required = false) String lastName
 ) {
         String accept = request.getHeader("Accept");
         if (accept != null && accept.contains("application/json")) {
@@ -107,7 +101,7 @@ public class UsersApiController implements UsersApi {
     }
 
     public ResponseEntity<Void> updateUser(@ApiParam(value = "User id to get from the database",required=true) @PathVariable("userId") Integer userId
-,@ApiParam(value = ""  )  @Valid @RequestBody RegistrationUser body
+,@ApiParam(value = ""  )  @Valid @RequestBody User body
 ) {
         String accept = request.getHeader("Accept");
         return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
