+++ src/main/java/io/swagger/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package io.swagger.repository;
+
+import io.swagger.model.Transaction;
+import io.swagger.model.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Integer> {
+
+}
