+++ src/main/java/io/swagger/service/UserService.java
@@ -1,0 +1,15 @@
+package io.swagger.service;
+
+import io.swagger.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    public UserService() {
+    }
+}
